Update time type from Integer to Long in model
Update time type from Integer to Long in model to correctly receive time.

+++ app/src/main/java/com/mck/weathernow/model/CurrentWeatherData.java
@@ -13,7 +13,7 @@ public class CurrentWeatherData {
     public Clouds clouds;
     public Rain rain;
     public Snow snow;
-    public Integer dt; // date in unix
+    public Long dt; // date in unix
     public Sys sys;
     public Integer id; // City id
     public String name; // City name
